requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
crispy-bootstrap5==0.7
Django==4.0.5
django-autocomplete-light==3.9.4
django-crispy-forms==2.0
Pillow==9.4.0
psycopg2-binary==2.9.5
six==1.16.0
sqlparse==0.4.2
